deliverables nanopore
